# stock_service/chain_layer/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CompanyX Inventory API",
    "version": "1.0.0",
    "description": "API for managing inventory for CompanyX"
  },
  "servers": [
    {
      "url": "https://145d1151ce28c6.lhr.life"
    }
  ],
  "paths": {
    "/overview": {
      "get": {
        "summary": "Get inventory overview",
        "operationId": "getInventoryOverview",
        "responses": {
          "200": {
            "description": "Successfully retrieved inventory summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search for an item",
        "operationId": "searchItem",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/price": {
      "put": {
        "summary": "Change price of an item",
        "operationId": "changePrice",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "option": {
                    "type": "string"
                  },
                  "item_id": {
                    "type": "string"
                  },
                  "change_value": {
                    "type": "number"
                  },
                  "is_percentage": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price changed successfully"
          }
        }
      }
    },
    "/movement-log": {
      "get": {
        "summary": "Retrieve stock movement logs",
        "operationId": "getMovements",
        "responses": {
          "200": {
            "description": "Successfully retrieved stock movement logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Log a stock movement",
        "operationId": "logMovement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "movement": {
                    "type": "object",
                    "properties": {
                      "item_id": {
                        "type": "string"
                      },
                      "quantity": {
                        "type": "integer"
                      },
                      "movement_type": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock movement logged successfully"
          }
        }
      }
    },
    "/stock-valuation": {
      "get": {
        "summary": "Retrieve stock valuation",
        "operationId": "getValuation",
        "parameters": [
          {
            "in": "query",
            "name": "num_items",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved stock valuation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/stock": {
      "post": {
        "summary": "Manage stock items (Add, Update, Delete)",
        "operationId": "stockManagement",
        "requestBody": {
          "description": "Stock management query",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "object",
                    "properties": {
                      "option": {
                        "type": "string",
                        "enum": ["add", "update", "delete"]
                      },
                      "item": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "SKU": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "cost": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "price": {
                            "type": "integer",
                            "minimum": 0
                          }
                        },
                        "required": ["name", "SKU", "description", "cost", "price"]
                      },
                      "new_level": {
                        "type": "integer"
                      },
                      "id": {
                        "type": "string"
                      }
                    },
                    "required": ["option"]
                  }
                },
                "required": ["query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock successfully managed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
    
    }
  }